Ticket #48048 - Fix coverity issues - 2015/2/24
Description: Due to the changes made on NSPRIu64 and NSPRI64 along with
this fix:
> Coverity defect 13075 - Explicit null dereferenced (FORWARD_NULL)
> NSPRIu64 and NSPRI64 is defined based upon the architecture and compiler.
casting PRInt64 to (long long int) in the log print is no longer needed.
Got rid of them.

Reviewed and advised by rmeggins@redhat.com (Thank you, Rich!!)

https://fedorahosted.org/389/ticket/48048
(cherry picked from commit 8f1eef1b53a88e5d42e3313eae5fe63b9a7dd99b)

+++ ldap/servers/slapd/log.c
@@ -2293,7 +2293,7 @@ log__open_accesslogfile(int logfile_state, int locked)
 		log_convert_time (logp->l_ctime, tbuf, 1 /*short*/);
 		PR_snprintf(buffer, sizeof(buffer), "LOGINFO:%s%s.%s (%lu) (%"
 			NSPRI64 "d)\n", PREVLOGFILE, loginfo.log_access_file, tbuf, 
-			logp->l_ctime, (long long int)logp->l_size);
+			logp->l_ctime, logp->l_size);
 		LOG_WRITE(fpinfo, buffer, strlen(buffer), 0);
 		logp = logp->l_next;
 	}
@@ -3906,7 +3906,7 @@ log__open_errorlogfile(int logfile_state, int locked)
 		log_convert_time (logp->l_ctime, tbuf, 1 /*short */);
 		PR_snprintf(buffer, sizeof(buffer), "LOGINFO:%s%s.%s (%lu) (%" 
 			NSPRI64 "d)\n", PREVLOGFILE, loginfo.log_error_file, tbuf,
-			logp->l_ctime, (long long int)logp->l_size);
+			logp->l_ctime, logp->l_size);
 		LOG_WRITE(fpinfo, buffer, strlen(buffer), 0);
 		logp = logp->l_next;
 	}
@@ -4033,7 +4033,7 @@ log__open_auditlogfile(int logfile_state, int locked)
 		log_convert_time (logp->l_ctime, tbuf, 1 /*short */);	
 		PR_snprintf(buffer, sizeof(buffer), "LOGINFO:%s%s.%s (%lu) (%"
 			NSPRI64 "d)\n", PREVLOGFILE, loginfo.log_audit_file, tbuf, 
-			logp->l_ctime, (long long int)logp->l_size);
+			logp->l_ctime, logp->l_size);
 		LOG_WRITE(fpinfo, buffer, strlen(buffer), 0);
 		logp = logp->l_next;
 	}
